pyproject.toml
[tool.poetry]
name = "tensaku"
version = "0.1.0"
description = ""
authors = ["Shohei Jinno <shohei.jinno@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
openai = "^1.55.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
